add TemperatureFader class
TemperatureFader is an optional module that will automatically modulate
a PWM output based on the temperature range. It is very similar to
TemperatureSwitch,
 but you specify a transition temperature range and an output PWM
range.  Temperature values less than the transition range will result
in the output being turned off, values above the range will
result in the output being set to the max value in the output PWM range.